Pass data from sysmon3 through sm3_config to sm3_temps.
Originally the udp class ways instantiated in both sysmon3 and
sm3_temps.  In some cases the instances were interfering and causing
a timout to kill the process. Since sm3_temps only needs to read
the interface and sensor names, we can just pass it from sysmon3.

In this case we need to go through the intermediate sm3_config
class.