Enhance network timeout hangling.
If the udp request fails because of network problems or if the
server is not responding, add the capability to retry the request
for data,

If it's a simple transient network delay, a retry will probably work.

If there is a problem with sysmond, restarting it on the server
before selecting retry may also work.