Use multiple configuration files.
Originally I was trying to use one configuration file for all running instances
of sysmon3.  I figured out that I can more easily set the filename of the
config file so I changed it to a common file, hosts.conf, to get things started
and then a specific <server>.conf file for each instance.

The contents of the configuration files are now much simpler.

All the configuration files are located at the directory
$HOME/.config/LinuxFromScratch/

Now the initial setting/reading of the target server is accessed directly from
the setup.cpp and sysmon3.cpp files, but all other configureation is done
through the SM_Config class.